+++ test/unit/tc_values.rb
@@ -92,7 +92,7 @@ class Test_Climate_values_constraints < Test::Unit::TestCase
 		stdout	=	StringIO.new
 		stderr	=	StringIO.new
 
-		climate = LibCLImate::Climate.new do |cl|
+		climate = LibCLImate::Climate.new(value_attributes: true) do |cl|
 
 			cl.exit_on_missing	=	false
 
@@ -129,6 +129,67 @@ class Test_Climate_values_constraints < Test::Unit::TestCase
 		assert_equal 2, r.values.size
 		assert_empty stdout.string
 		assert_empty stderr.string
+		assert r.respond_to? :input_path
+		assert r.respond_to? :output_path
+
+		stdout.string = ''
+		stderr.string = ''
+		r = climate.run [ 'value-1', 'value-2', 'value-3' ]
+		assert_equal 3, r.values.size
+		assert_empty stdout.string
+		assert_not_empty stderr.string
+		assert_match /wrong number of values.*3 given.*2 required.*/, stderr.string
+	end
+
+	def test_constrain_with_integer_and_names_2
+
+		stdout	=	StringIO.new
+		stderr	=	StringIO.new
+
+		climate = LibCLImate::Climate.new(value_attributes: true) do |cl|
+
+			cl.exit_on_missing	=	false
+
+			cl.stdout	=	stdout
+			cl.stderr	=	stderr
+
+			cl.constrain_values = 1..2
+			cl.value_names = [
+
+				'input-path',
+				'output-path',
+			]
+		end
+
+		stdout.string = ''
+		stderr.string = ''
+		r = climate.run [ ]
+		assert_equal 0, r.values.size
+		assert_empty stdout.string
+		assert_not_empty stderr.string
+		assert_match /input-path not specified.*#{climate.usage_help_suffix}/, stderr.string
+
+		stdout.string = ''
+		stderr.string = ''
+		r = climate.run [ 'value-1' ]
+		assert_equal 1, r.values.size
+		assert_empty stdout.string
+		assert_empty stderr.string
+		assert r.respond_to? :input_path
+		assert r.respond_to? :output_path
+		assert_equal 'value-1', r.input_path
+		  assert_nil r.output_path
+
+		stdout.string = ''
+		stderr.string = ''
+		r = climate.run [ 'value-1', 'value-2' ]
+		assert_equal 2, r.values.size
+		assert_empty stdout.string
+		assert_empty stderr.string
+		assert r.respond_to? :input_path
+		assert r.respond_to? :output_path
+		assert_equal 'value-1', r.input_path
+		assert_equal 'value-2', r.output_path
 
 		stdout.string = ''
 		stderr.string = ''